Fixed long long typedef in finwire parser

+++ include/FinwireParser.h
@@ -31,9 +31,9 @@ private:
         FIN = 0, CMP = 1, SEC = 2
     };
     char const _search[3][4] = {"FIN", "CMP", "SEC"};
-    ull const _FINLengths[18] = {15, 3, 4, 1, 8, 8, 17, 17, 12, 12, 12, 17, 17, 17, 13, 13, 60, 0};
-    ull const _CMPLengths[17] = {15, 3, 60, 10, 4, 2, 4, 8, 80, 80, 12, 25, 20, 24, 46, 150, 0};
-    ull const _SECLengths[13] = {15, 3, 15, 6, 4, 70, 6, 13, 8, 8, 12, 60, 0};
+    int const _FINLengths[18] = {15, 3, 4, 1, 8, 8, 17, 17, 12, 12, 12, 17, 17, 17, 13, 13, 60, 0};
+    int const _CMPLengths[17] = {15, 3, 60, 10, 4, 2, 4, 8, 80, 80, 12, 25, 20, 24, 46, 150, 0};
+    int const _SECLengths[13] = {15, 3, 15, 6, 4, 70, 6, 13, 8, 8, 12, 60, 0};
     af::array _data;
     af::array _indexer;
